Viewport fixes

+++ lib/tlQtWidget/TimelineWidget.h
@@ -99,9 +99,9 @@ namespace tl
                 const std::shared_ptr<ui::IWidget>&,
                 const math::BBox2i&);
 
-            float _toUI(float) const;
+            int _toUI(int) const;
             math::Vector2i _toUI(const math::Vector2i&) const;
-            float _fromUI(float) const;
+            int _fromUI(int) const;
             math::Vector2i _fromUI(const math::Vector2i&) const;
 
             TLRENDER_PRIVATE();
